requirements.txt
aiohappyeyeballs==2.4.3 ; python_version >= "3.9" and python_version < "4.0"
aiohttp==3.11.6 ; python_version >= "3.9" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.9" and python_version < "4.0"
aiosqlite==0.20.0 ; python_version >= "3.9" and python_version < "4.0"
alembic==1.14.0 ; python_version >= "3.9" and python_version < "4.0"
anyio==4.0.0 ; python_version >= "3.9" and python_version < "4.0"
async-timeout==5.0.1 ; python_version >= "3.9" and python_version < "3.11"
asyncpg==0.30.0 ; python_version >= "3.9" and python_version < "4.0"
attrs==24.2.0 ; python_version >= "3.9" and python_version < "4.0"
cachetools==5.5.0 ; python_version >= "3.9" and python_version < "4.0"
certifi==2023.7.22 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.4.0 ; python_version >= "3.9" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
exceptiongroup==1.1.3 ; python_version >= "3.9" and python_version < "3.11"
frozenlist==1.5.0 ; python_version >= "3.9" and python_version < "4.0"
google-api-core==2.23.0 ; python_version >= "3.9" and python_version < "4.0"
google-api-python-client==2.153.0 ; python_version >= "3.9" and python_version < "4.0"
google-auth-httplib2==0.2.0 ; python_version >= "3.9" and python_version < "4.0"
google-auth-oauthlib==1.2.1 ; python_version >= "3.9" and python_version < "4.0"
google-auth==2.36.0 ; python_version >= "3.9" and python_version < "4.0"
googleapis-common-protos==1.66.0 ; python_version >= "3.9" and python_version < "4.0"
greenlet==3.1.1 ; python_version < "3.13" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.9"
h11==0.14.0 ; python_version >= "3.9" and python_version < "4.0"
httpcore==0.17.3 ; python_version >= "3.9" and python_version < "4.0"
httplib2==0.22.0 ; python_version >= "3.9" and python_version < "4.0"
httpx==0.24.1 ; python_version >= "3.9" and python_version < "4.0"
idna==3.4 ; python_version >= "3.9" and python_version < "4.0"
mako==1.3.6 ; python_version >= "3.9" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.9" and python_version < "4.0"
multidict==6.1.0 ; python_version >= "3.9" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.9" and python_version < "4.0"
propcache==0.2.0 ; python_version >= "3.9" and python_version < "4.0"
proto-plus==1.25.0 ; python_version >= "3.9" and python_version < "4.0"
protobuf==5.28.3 ; python_version >= "3.9" and python_version < "4.0"
psycopg2-binary==2.9.10 ; python_version >= "3.9" and python_version < "4.0"
pyasn1-modules==0.4.1 ; python_version >= "3.9" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.9" and python_version < "4.0"
pyparsing==3.2.0 ; python_version >= "3.9" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.9" and python_version < "4.0"
python-telegram-bot==20.5 ; python_version >= "3.9" and python_version < "4.0"
requests-oauthlib==2.0.0 ; python_version >= "3.9" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.9" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.9" and python_version < "4"
sniffio==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
sqlalchemy==2.0.36 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.7.1 ; python_version >= "3.9" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.9" and python_version < "4.0"
urllib3==2.2.3 ; python_version >= "3.9" and python_version < "4.0"
yarl==1.17.2 ; python_version >= "3.9" and python_version < "4.0"
